somewhat of an image
Row timing is consistent with the ~20ms timing observed on the
Saleae.  Vsync interrupt timing is also consistent.  However, only 20
rows are captured at the point the the vsync interrupt occurs.

So it appears the time required to send a single row is perhaps the
time required to capture 2 more rows.  Thus we're only seeing 1/3 of
the rows.