corrected some bugs, but...
unclear why the previous commit actually produced a recognizable
image.  the offsets in the for loop were actually combining data from
neighboring pixels.  it was also iterating beyond the end of the
captured pixels, and this explains the uniform gray rectangle next to
the image.

after correcting these mistakes, and trunctating the first and last 25
pixels per row, it captures exactly 100 pixels per row.  but the image
it produces isn't recognizable.